treewide: Merge all @MissMinji_bot changes into master
This is done to make a commits log clean.

-- commits contributors --
* `shrp` function in android.py is done by corsicanu in GitHub.
-> https://github.com/corsicanu/tgbot
* Rewritten paginate_modules to @MissRose_bot style is done by dank-del in GitHub.
-> https://github.com/Dank-del/EnterpriseALRobot

+++ tg_bot/modules/sql/global_kicks_sql.py
@@ -1,13 +1,11 @@
-import threading
-
-from sqlalchemy import Column, UnicodeText, Integer, String, Boolean
+from sqlalchemy import Column, UnicodeText, BigInteger, Integer
 
 from tg_bot.modules.sql import BASE, SESSION
 
 
 class GloballyKickedUsers(BASE):
     __tablename__ = "gkicks"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     name = Column(UnicodeText, nullable=False)
     times = Column(Integer)
 
@@ -17,11 +15,7 @@ class GloballyKickedUsers(BASE):
         self.times = times
 
     def to_dict(self):
-        return {
-            "user_id": self.user_id,
-            "name": self.name,
-            "times": self.times
-        }
+        return {"user_id": self.user_id, "name": self.name, "times": self.times}
 
 
 GloballyKickedUsers.__table__.create(checkfirst=True)
@@ -69,9 +63,6 @@ def get_times(user_id):
 def __load_gkick_userid_list():
     global GKICK_LIST
     try:
-        GKICK_LIST = {
-            x.user_id
-            for x in SESSION.query(GloballyKickedUsers).all()
-        }
+        GKICK_LIST = {x.user_id for x in SESSION.query(GloballyKickedUsers).all()}
     finally:
         SESSION.close()
